changed headers sca and swMIp to amd and ICI, respectively in files final4.csv
in AA min distace made several changes now it reads a file a returns a file with the minimal distance for AB hetodimer's interface amino acids and read this file to return the minimum value encountered.

In MIp chaged the length of the variable form 6 to 5 position in the file's name

IN PII retriever added some variables at the star for to ease the use of the program

Added pii test.csv as it is needed by Pii retriever to work.